geon_azure_voice_agent/requirements.txt

fastapi
uvicorn[standard]
httpx
websockets
python-dotenv
